Update image reference syntax and enhance UI refresh mechanisms
- Updated image reference syntax in the API and components from `@image_name` to `{{image_name}}` for consistency across the application.
- Enhanced ClientSTMDemo to include a `chatKey` state for forcing remounts on load/import/clear operations, improving UI responsiveness.
- Introduced a `handleFullRefresh` callback in ClientSTMDemo to trigger UI updates after significant state changes.
- Refactored STMEditor to support a new `onFullRefresh` prop, allowing parent components to refresh the UI effectively.
- Cleaned up unused code and improved overall component structure for better maintainability.
