+++ examples/nextjs_demo/src/components/STMEditor.tsx
@@ -2,6 +2,7 @@
 
 import { useState, useCallback, useRef, useEffect } from 'react';
 import { mindcache } from 'mindcache';
+import STMMenu from './STMMenu';
 
 // Type definitions
 interface ToolSchema {
@@ -10,9 +11,10 @@ interface ToolSchema {
 
 interface STMEditorProps {
   onSTMChange?: () => void;
+  onFullRefresh?: () => void; // Called on load/import/clear to refresh chat & workflows
 }
 
-export default function STMEditor({ onSTMChange }: STMEditorProps) {
+export default function STMEditor({ onSTMChange, onFullRefresh }: STMEditorProps) {
   const mindcacheRef = useRef(mindcache);
   const [stmState, setSTMState] = useState(mindcacheRef.current.getAll());
   const [editingKey, setEditingKey] = useState<string | null>(null);
@@ -45,33 +47,6 @@ export default function STMEditor({ onSTMChange }: STMEditorProps) {
     return () => mindcacheRef.current.unsubscribeFromAll(updateSTMState);
   }, [updateSTMState]);
 
-  // Global keyboard shortcuts for terminal commands
-  useEffect(() => {
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if (e.ctrlKey || e.metaKey) {
-        switch (e.key) {
-          case 's':
-            e.preventDefault();
-            saveSTM();
-            break;
-          case 'l':
-            e.preventDefault();
-            loadSTM();
-            break;
-          case 'k':
-            e.preventDefault();
-            if (confirm('Clear STM? This will restore default values.')) {
-              clearSTM();
-            }
-            break;
-        }
-      }
-    };
-
-    window.addEventListener('keydown', handleKeyDown);
-    return () => window.removeEventListener('keydown', handleKeyDown);
-  }, []);
-
   // Generate tool schemas (without execute functions) for display
   function getToolSchemas(): Record<string, ToolSchema> {
     const tools = mindcacheRef.current.get_aisdk_tools();
@@ -86,13 +61,6 @@ export default function STMEditor({ onSTMChange }: STMEditorProps) {
 
     return schemas;
   }
-
-  // Add a new STM key
-  const addSTMKey = (key: string) => {
-    if (key && !mindcacheRef.current.has(key)) {
-      mindcacheRef.current.set_value(key, '');
-    }
-  };
 
   // Handle file upload
   const handleFileUpload = async (key: string, file: File) => {
@@ -265,153 +233,18 @@ export default function STMEditor({ onSTMChange }: STMEditorProps) {
     }
   };
 
-  // Save STM to localStorage
-  const saveSTM = () => {
-    try {
-      const serialized = mindcacheRef.current.toJSON();
-      localStorage.setItem('mindcache_stm', serialized);
-      console.log('✅ STM saved to localStorage');
-    } catch (error) {
-      console.error('❌ Failed to save STM:', error);
+  // Handle full refresh (load/import/clear operations)
+  const handleFullRefresh = useCallback(() => {
+    updateSTMState(); // Update local state first
+    if (onFullRefresh) {
+      onFullRefresh(); // Then trigger parent refresh (chat reset, workflows)
     }
-  };
-
-  // Load STM from localStorage
-  const loadSTM = () => {
-    try {
-      const saved = localStorage.getItem('mindcache_stm');
-      if (saved) {
-        mindcacheRef.current.fromJSON(saved);
-        console.log('✅ STM loaded from localStorage');
-      } else {
-        console.log('ℹ️ No saved STM found');
-      }
-    } catch (error) {
-      console.error('❌ Failed to load STM:', error);
-    }
-  };
-
-  // Clear STM
-  const clearSTM = () => {
-    mindcacheRef.current.clear();
-    console.log('🗑️ STM cleared');
-  };
-
-  // Export STM to markdown file
-  const exportSTM = () => {
-    try {
-      const markdown = mindcacheRef.current.toMarkdown();
-      const blob = new Blob([markdown], { type: 'text/markdown' });
-      const url = URL.createObjectURL(blob);
-      const a = document.createElement('a');
-      a.href = url;
-      a.download = `mindcache-export-${new Date().toISOString().split('T')[0]}.md`;
-      document.body.appendChild(a);
-      a.click();
-      document.body.removeChild(a);
-      URL.revokeObjectURL(url);
-      console.log('✅ STM exported to markdown');
-    } catch (error) {
-      console.error('❌ Failed to export STM:', error);
-    }
-  };
-
-  // Import STM from markdown file
-  const importSTM = () => {
-    try {
-      const input = document.createElement('input');
-      input.type = 'file';
-      input.accept = '.md,.markdown,text/markdown';
-      input.onchange = (e) => {
-        const file = (e.target as HTMLInputElement).files?.[0];
-        if (file) {
-          const reader = new FileReader();
-          reader.onload = (event) => {
-            try {
-              const markdown = event.target?.result as string;
-              mindcacheRef.current.fromMarkdown(markdown);
-              updateSTMState(); // Refresh UI and notify listeners
-              console.log('✅ STM imported from markdown');
-            } catch (error) {
-              console.error('❌ Failed to parse markdown:', error);
-              alert('Failed to import markdown file. Please check the file format.');
-            }
-          };
-          reader.onerror = () => {
-            console.error('❌ Failed to read file');
-            alert('Failed to read file.');
-          };
-          reader.readAsText(file);
-        }
-      };
-      input.click();
-    } catch (error) {
-      console.error('❌ Failed to import STM:', error);
-    }
-  };
+  }, [onFullRefresh, updateSTMState]);
 
   return (
     <div className="flex-1 flex flex-col pl-1 min-h-0">
-      {/* Terminal Commands - Fixed Header */}
-      <div className="border border-green-400 rounded-t p-4 border-b-0 font-mono text-sm flex-shrink-0">
-        <div className="flex space-x-4 mb-2">
-          <div 
-            className="text-green-400 cursor-pointer hover:text-green-300 transition-colors"
-            onClick={() => {
-              const key = prompt('Enter new STM key:');
-              if (key && key.trim()) {
-                addSTMKey(key.trim());
-              }
-            }}
-            title="Add new STM key"
-          >
-            Add Key
-          </div>
-          <div 
-            className="text-green-400 cursor-pointer hover:text-green-300 transition-colors"
-            onClick={loadSTM}
-            title="Load STM from localStorage (Ctrl+L)"
-          >
-            Load
-          </div>
-          <div 
-            className="text-green-400 cursor-pointer hover:text-green-300 transition-colors"
-            onClick={saveSTM}
-            title="Save STM to localStorage (Ctrl+S)"
-          >
-            Save
-          </div>
-          <div 
-            className="text-green-400 cursor-pointer hover:text-green-300 transition-colors"
-            onClick={() => {
-              if (confirm('Clear STM? This will restore default values.')) {
-                clearSTM();
-              }
-            }}
-            title="Clear STM - keeps defaults (Ctrl+K)"
-          >
-            Clear
-          </div>
-          <div 
-            className="text-green-400 cursor-pointer hover:text-green-300 transition-colors"
-            onClick={exportSTM}
-            title="Export STM to markdown file"
-          >
-            Export
-          </div>
-          <div 
-            className="text-green-400 cursor-pointer hover:text-green-300 transition-colors"
-            onClick={importSTM}
-            title="Import STM from markdown file"
-          >
-            Import
-          </div>
-        </div>
-        <div className="text-xs text-gray-500 mb-4">
-          Auto-loads on page refresh • Ctrl+S/L/K shortcuts • Export/Import as markdown
-        </div>
-        <div className="border-b border-green-400"></div>
-      </div>
+      {/* STM Menu - Fixed Header */}
+      <STMMenu onRefresh={handleFullRefresh} />
 
       {/* STM Content - Scrollable */}
       <div className="flex-1 border border-green-400 rounded-b p-4 overflow-y-auto min-h-0 border-t-0">
